backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.5
tzdata==2026.1
